feat: same source copy implemented
