+++ src/components/main/actionsPanel/workspaceTreeView/WorkspaceTreeView.tsx
@@ -441,7 +441,7 @@ export default function WorkspaceTreeView() {
             // update the invalid file nodes
             dispatch(removeInvalidFileNodes([node.uid]));
           },
-          onRenameItem: (item, name) => {
+          onRenameItem: (item, entityName) => {
             // cb_renameNode(item, name);
             const node = item.data as TNode;
             const nodeData = node.data as TFileNodeData;
@@ -453,14 +453,14 @@ export default function WorkspaceTreeView() {
               const parentNode = fileTree[node.parentUid!];
               if (!parentNode) return;
 
-              const newUid = _path.join(parentNode.uid, name);
+              const newUid = _path.join(parentNode.uid, entityName);
               const isFolder = !node.isEntity;
               if (isFolder) {
-                rnbw.files.createFolder({ name });
+                rnbw.files.createFolder({ entityName });
               } else {
                 const ext = nodeData.ext;
                 rnbw.files.createFile({
-                  name: name,
+                  entityName,
                   extension: ext,
                 });
